QuestionAnswering/QA_ICL/deployment/fastapi_app/requirements.txt
fastapi
uvicorn[standard]
redis
SPARQLWrapper
openai
tritonclient[grpc]
numpy
jinja2
pydantic-settings
transformers
tiktoken
shapely